+++ pendulum/tz/__init__.py
@@ -1,7 +1,12 @@
 # -*- coding: utf-8 -*-
 
+import pytzdata
+
 from .timezone import Timezone, FixedTimezone, UTC
 from .local_timezone import LocalTimezone
+
+
+timezones = pytzdata.timezones
 
 
 def timezone(name):