+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Added support for the alternative formatter's tokens in `from_format()`
 - Added a `timezones` module attribute to expose available timezones.
 
+### Changed
+
+- Dopped support for Python 3.2 and 3.3.
+
 
 ## [1.2.5] - 2017-09-04